+++ src/EntityFramework.Relational.Design/Metadata/Internal/ScaffoldingAnnotationNames.cs
@@ -1,7 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-namespace Microsoft.Data.Entity.Scaffolding.Metadata
+namespace Microsoft.Data.Entity.Scaffolding.Metadata.Internal
 {
     public class ScaffoldingAnnotationNames
     {
