+++ src/EntityFramework.Relational/Metadata/RelationalAnnotations.cs
@@ -4,6 +4,7 @@
 using System.Diagnostics;
 using JetBrains.Annotations;
 using Microsoft.Data.Entity.Infrastructure;
+using Microsoft.Data.Entity.Metadata.Internal;
 using Microsoft.Data.Entity.Utilities;
 
 namespace Microsoft.Data.Entity.Metadata
